Employ the new algorithm to compute the polynomials
Had to change some stuff with respect to the string formatting now that
we're working with floating point coefficients and since there's now an
option to return the coefficient vector there are a few conditions which
prevent printing a long output for high orders.

The LaTeX title is now also suppressed if it's too long (checked against
a maximum order rather than string length) and as long as we're doing
this then there's another new option to provide the LaTeX string.

The function also now returns the figure.