Subtract the piston and tilt terms when truncating the plot title
If we're only going to be presenting the most significant terms when
there are a large number of them we should only retain those related
to differences in the wavefront curvature.

Also, the sign on the leading coefficient in this scenario has been
fixed.

The string formatting for the wavefront plot title has become somewhat
difficult to understand and there's probably a better way to implement
this, but this will have to do for now.