Add option to return wavefront error function, closure optimizations
There was already a way to return the Zernike polynomial function by
using the fitting mode so this provides a way to extract the model
as a closure which could be useful if you want to plot it using a
different package or inspect the fitted values over a region of
the pupil in finer detail.

Type annotations and let blocks have also been added regarding these two
closures for performance as the compiler apparently has a difficult time
in these sorts of situations.