Revert the closure "optimizations" from the previous commit
So yeah after some cursory benchmarking with BenchmarkTools it seems
there's no benefit and if anything performance has decreased. Either the
compiler has figured this stuff out or there's something I don't
understand (most likely). So whatever, I'll investigate this some
other time.

In any case, the biggest performance issue with the package right now is
the type instability which at the moment I don't know how to fix.
The problem is the main Z function is used for a variety of things;
right now that function:

- computes the polynomials
- formats the output strings
- calls the plotting function
- provides the functions for the wavefront error fitting
- returns the coefficients and the LaTeX formatted string if desired

and since it's accepting keyword arguments and its return type varies
it's a mess. Usually I guess you would define some custom tailored types
using structs and use that to define individual methods for each, but
I'm currently not knowledgeable enough to implement this elegantly;
at least I don't see a way without redefining the main body of the
function each time.

The issue with the type instability here is that the function is used in
a loop at W in order to create a vector of Zernike polynomial functions
up to the specified max order so this should be as fast as possible and
ideally we'd have a fixed return type for the Z function.

That being said, maybe the compiler can infer the type using constant
propagation (even though fit is not a positional argument); at least I
don't see a difference after probing.

Anyways, there are a few more features and things to implement and the
speed is good enough right now for a reasonable number of Zernike terms;
the polynomial computation is fast enough with the new algorithm, it's
really only the construction of the wavefront error model which could be
improved for high orders.