Prevent aliasing by dynamically determining the sample rate
For the angular component determining the Nyquist rate is trivial, but
for the radial polynomials some analysis is required. If you extend them
periodically and approximate using a Fourier series there emerges a
pattern where the linear frequency increases by a quarter for each
increase in n, which yields pi * n / 2 for the bandwidth.

The maximum length has been set to 1000 which bounds the number of
elements in the matrix to a million for performance reasons. Aliasing
should not be an issue until polynomial degrees past 300 or so in this
case and for lower degrees since we're scaling by two orders of
magnitude the plots should appear less jagged compared to before.