Redesign in an attempt to address the type instability
Hopefully I've done this gracefully enough. I tried to avoid code
duplication and tried to structure everything as logically as I possibly
could. What I don't really like is having to pack and unpack the local
variables from Zf, but it doesn't really make sense to have the string
formatting within that function and when the wavefront error fitting is
done there's no need for it; the string formatting depends upon a heavy
number of parameters which is why I decided to do it that way.

The return types for the most computationally expensive functions appear
to be stable at least.

I benchmarked the closure hacks again and still no benefit, so I'm
leaving it like this for now.

The wavefront error modelling is still very slow for a large number of
points across the disk, but this will hopefully be solved in the next
commit when I add options to control the precision.