Manually create the workhorse functions explicitly using composite types
This essentially does the same thing, but it's cleaner. Especially since
we're working with closures it makes sense to specify these as callable
functors and use our own types for them. It also makes it easier to work
with and access the fields (e.g.: Znm.R compared to Znm.R.contents)
which is going to be important for subsequent commits which will need to
access the radial polynomials specifically.

Now, a couple of notes regarding what could be improved:

First, the separable components of the Zernike function are not
annotated using concrete types. This could possibly be improved with
parameterization, but there could be a better solution altogether than
storing functions inside of a struct.

Second, there might be too many singleton types for the options.
Maybe some encompassing Extract() constructor which yields both the
coefficients and the LaTeX string would be better.