Add a pupil scaling algorithm
This yields the Zernike expansion coefficients on stopping down the lens
which clips the wavefront at the edges producing a different relation in
terms of Zernike polynomials.

Now, because the emergent wavefront remains invariant over the region of
the contracted exit pupil we could in theory use the wavefront model
function which was built using the original fitted coefficients then
simply sample that wavefront and refit, but this process is slower,
prone to error if the sampling rate (which might need to be quite high
for high orders) isn't correctly chosen, and is not as elegant as the
approach taken.

Note that in Janssen & Dirksen's paper the parameter they use for the
coefficients, α, includes the normalization factor N so this had to be
accounted for.