Integrate the scaling algorithm with the rest of the package
I can't decide whether this is bordering on spaghetti code or if it
makes sense to structure it this way. The trade-off is that there's good
reuse, little to no code duplication, and the functions aren't huge, but
the drawback is that the flow is much more difficult to understand.

And it doesn't help that descriptive names aren't being used for these
functions either; I'm trying to keep with the general theme for the
package (which is already contrary to Julia conventions) and stay
consistent, but as complexity increases the code base becomes more
abstruse. Besides, I can't really find good names for these functions
anyway.

Basically, there's three levels/tiers to the structure:

* Interface -->
  * Construction -->
    * Synthesis.

And this is true for both the Zernike part and the wavefront error
aspect.

The interface accepts the initial input, the construction phase builds
the required (ρ, θ) -> functions Zmn and ΔW, and the synthesis stage
assembles the necessary output.

For the Zernikes we have:
in (with options) --> synth calls for construction --> assembly --> out

For the wavefront errors:
in (with options) --> construction --> filter --> synth assembly --> out

This allows for modularity and keeps the construction functions light
(the Zernike one gets called in loops) and doing it this way makes it
easy to integrate new features if some of the same functionality is
desired on output (as is the case here with the plot, coefficients, and
metrics under scaling operations). Note also that this way the functions
are available to be called manually as well if so desired.

For the wavefront error there's an additional filtering step because we
only want to use the Zernike polynomials with non-zero expansion
coefficients so the precision option is used for that.

Either way, the program works, it's type stable, and it's fast for the
most part. But yeah, readability is an issue. Maybe the structure's fine
maybe it's really just the function names.