Have Z(j) accept kwargs, explicitly state them in the other methods
The former was overlooked, the latter makes querying the methods yield
more information about the expected arguments and since there's only one
option with a keyword argument there's no need to condense.

This is still far from ideal, however, as we mix options using
positional arguments and keyword arguments (and "opts", "options" is
ambiguous and confusing if you don't look at the documentation which
doesn't exist yet). This is done because we have conditional return
values in order to avoid dumping extra stuff unless requested and
different methods are used for this in order to have type stability.

Also, a crash was fixed if you specified incompatible options
(for example, Fit() and scale) so now there's an extra check to make
sure the right method is called.