Redesign #2: custom composite types for output; overload show, iterate
This merges the interface and synthesis processes for the Zernike
computation while the synth function is retained for the wavefront error
computation as both the initial fitting feature and the scaling feature
utilize it, but under a different context.

For the Zernike polynomials it's better to return everything and just
clean up the output by extending show rather having multiple singleton
types corresponding to different return values. Here iterate is also
extended to allow for the choice to destructure the output without using
property destructuring.

The functor types have had their display cleaned up as well, with the
wavefront error now also containing the tuple of indices for each
polynomial used. While this is somewhat redundant (as the vector of
expansion coefficients are already stored in one of the fields) this
avoids having to do extra work on show or when calling the functor.

Now, the naming is somewhat confusing for the expansion coefficients
as several names are used for different versions of them. This is really
only a problem for anyone reading the source and trying to make sense of
this mess as the interface doesn't really expose these potential
ambiguities unless you drill into the fields at which point it becomes
clear what is what.

Basically three different versions of these coefficients are used:

* the full vector of coefficients as computed by linear least squares
* the filtered vector of coefficients post precision rounding
* the filtered coeffs stored in a vector of named tuples with the inds

And various names are used for these: v, av, a, i

With the first three remaining consistent outside of the struct with the
aforementioned three versions respectively. The `i` used for the
WavefrontError struct is just `a`, but `a` is already used as a field
and `a` here is really just av, which could be confusing.

inds isn't really a proper name for the former (as the tuples also store
the coefficients) and it appears to be more clear using `a` rather than
`av` for the calculation.

In any case, the naming is going to remain context dependent for
the time being.