Customize the display resolution, fontsize
On the one hand this is better as it avoids hardcoding the values and
instead personalizes it per device, on the other it might not work
across all platforms. This is generally preferrable, however, as it
provides consistency for various display resolutions and sizes.
Basically this means the window should look more or less the same as
the screen resolution and DPI are being used to determine the
presentation.