format_strings: use an input polynomial instead of a dictionary
So this simplifies a few things, removes the dependency on UnPack, and
allows for better type inference. Packing and unpacking local variables
like that is really ugly and in fact in this case unnecessary given that
almost all of the values needed are embedded within the polynomial;
the only two that aren't are extracted using inexpensive computations
and so we can do away with this business and have Zf return just the
polynomial.