Enable bounds checking in the two algorithms
So after some benchmarking there seems to be no difference and while
I've never encountered an issue and I'm pretty sure everything is
inbounds I'm enabling it. The inner loops are sometimes defined over
empty ranges and while everything seems correct it's better to be safe
especially if there's no impact on performance.