Implement the original radial coefficient formula using floats
Obviously this isn't used, but it's nice to have around for comparison
and so it's better to use a factorial method on floats numerically equal
to integers rather than the previous dance with ints.

Also this function shouldn't be polluting the Zernike kernel if it's
going to be commented out so I've moved it to the top level.