TransformAperture: compute the N & R matrices
So instead of utilizing a permutation matrix to switch the order
such that R is a block-diagonal matrix I've decided to utilize an
indexing scheme. Although this is non-standard, for me it's more
intuitive, but I can see this being somewhat confusing for anyone
reading through the source for the first time.

The paper I'm basing this algorithm on (Lundström & Unsbo 2007) does not
do it this way, but the MATLAB code they provide is somewhat difficult
for me to follow (it's been years since I've written MATLAB code) and
Julia allows for a more elegant (and possibly more efficient) approach.

That being said, I'm not going to try to benchmark their implementation
against mine; it's quite a well written paper that their mathematical
presentation of the method is clear enough that it makes custom
implementations straight-forward and this should be fast enough.

All that's left is to compute the transformation matrices and then
interface the functions.