TransformAperture: test the new algorithm, implement pupil scaling
From what I can see this is very similar in performance to the algorithm
found in ScaleAperture.jl. Some cursory benchmarking actually yields
better results for the complex matrix implementation both in terms of
speed and especially in terms of number of allocations; this is probably
due to the fact that the algorithm based on Janssen & Dirksen's formula
has to construct the Zernike Polynomial function and call the radial
component R(ε), possibly leading to some overhead.

The accuracy is practically identical (comparing it to an unscaled full
precision model function ΔW(ερ, θ)), with errors on the order of the
machine epsilon for floats.

ScaleAperture.jl will probably be relegated to unused comparison code,
given that TransformAperture hasn't even been polished or optimized yet,
and that it will encompass not just scaling, but also translation and
rotation; translation in this case is coupled to the scaling.