TransformAperture: construct the translation matrix
This requires a two-pass system as access to as of yet unmapped indices
is required to construct the translation matrix for each iteration. This
isn't much of a problem however as there isn't much of a performance
penalty for looping through twice since remap doesn't do much work. Note
that we can get away with not presetting the mapping for diagonal or
block-diagonal matrices like the pure scale matrix or the coefficient
matrix.

The pure scale matrix is kept because even though the translation matrix
covers scaling as well for zero shift it's a bit slower since it
requires another inner loop and extra computation. Once the rotation
matrix is included I will probably add a few extra conditions to ensure
these extra computations aren't carried out unnecessarily.