TransformAperture: transformation matrix splitting
This type of approach at bifurcation is really messy, but I've
benchmarked several different implementations and this type of
management yields the fastest results. The simplest alternative is of
course initializing and constructing all of the matrices resulting in
identity matrices for transformation values of zero. From what I can
tell the issue with this is that unnecessarily using complex matrices is
much slower (for example, just straight scaling only requires real
valued floats) and the computation behind the translation matrix is
somewhat expensive.

Now, there are two drawbacks to the current approach: namely that the
algorithm is not as clean as it could be due to all of the extra
conditions, and that Union types are used because the final
transformation and conversion matrices could be complex or not depending
on the type of transformation applied.

That being said, splitting this into separate methods for each
coordinate transformation seems to be more trouble than it's worth and
Julia seems to handle small Union types well. Only scaling is a third
faster than ScaleAperture while scaling + translation practically
matches it so for now this is good enough.