TransformAperture: create a separate function for translation
Okay so if we're going to have a two-pass system this makes a lot more
sense. To be specific, the coefficient matrix does not need a preset
mapping since for each iteration the m map is fixed and what is
essentially n′ (n-2s) only decreases (< n) from within the inner-most
loop while the upper-level nested loop increments n; this means that
there will always be a mapping available from a previous iteration.

In contrast, the translation matrix's loop varies both m′ and n′ so we
need to go through twice. But if we're going to do that instead of
having a separate remap function a better way is to have a dedicated
routine for the entire translation process since it's the real reason
why we need another loop in the first place.