Free up some cross-function dependent parameters
While this does require some extra (negligible) calls it improves the
structure in several ways. First, there's no need for the scaling
function to return what is essentially the length; this is just noise.
Next, it makes the translate function more independent in the sense that
it can be called in a more direct manner without having the length or
n_max on hand; this makes it especially useful for extracting the
translation matrix if you have the remap dictionary. The redundancy here
doesn't really matter.