Sum translation element terms corresponding to ⟨ρM′| coefficients
This does not appear to be explicitly necessary which is why I neglected
it to begin with as the corresponding terms are orthogonal in the sense
that m′, n′ pairs are unique for each full p, q loop resulting in an
injective mapping. I have not proven this in a sufficiently rigorous
manner, however, and working on the elliptical pupil transform has made
clear to me that this does in fact matter in the wider, more general
sense.

For the elliptical transform the (p, q) => (m′, n′=n) mapping is
non-injective which means that as the column is constructed for each
(m, n) pair identical row indices will sometimes be used to overwrite
existing values within the matrix.

So to be safe, and to present the implementation in a mathematically
correct way it is better to explicitly sum the elements. I checked and
this does not affect performance in any appreciable way.