Type stability for translation + elliptical transforms using 3 functions
This avoids Union return types at the expense of a little code pasting.
But again this way we do get to have separate functions for extracting
the transformation matrices for translation and elliptical pupils
while minimizing the number of loops.

The encompassing transform function still deals with Union types of
complex and non-complex matrices for the scaling, identity and complex
matrices for the other transformations. This should be fine, however,
since the return types are properly inferred and the Unions are small.

Having fixed types here is in fact slower as you have to instantiate
complex matrices which wind up as unity anyway after looping
excessively.

The dependency on LinearAlgebra is there for the identity matrix in
order to avoid conditionally setting the net transform matrix which
would require even more conditions and the number of permutations there
is annoying (2^n - 1 = 7).

The reason the algorithm is currently structured the way it is is for
parsimony, as chaining multiple transforms while having dedicated
functions for each would require multiple loops and increased
computation time. This is still necessary if a different order of
transformations is desired, but if you want:

scaling -> translation -> rotation -> elliptical

in that order then this seems to be the fastest way to do it using the
current scheme; here you only need two loops: the initial one for
N, R, and the remap and then the mapped loop for the more complicated
translational and elliptical transformations. Then we can take advantage
of the fact that the conversion matrix can be expressed as:

C = [N]⁻¹[R]⁻¹[ηₑ][ηᵣ][ηₜ][R][N] = ([R][N])⁻¹[η][R][N]

Anyway, I'm sure there's a better way to do all of this, but for now
this is fast enough and does what it needs to do.