Plot a log transform for high order Zernike polynomials
Without this quite a bit of detail is lost in the visualization for high
orders due to the large range of values involved. Makie's Axis3 does not
appear to have a log scale plotting attribute like Axis does so this
will have to do for now.

For wavefront errors this is not done as peak-to-valley errors large
enough to require a logarithmic presentation is not common as far as I
know, and besides when working with high order aberrations the focus
will be not so much on the plot, but rather on the data manipulation;
either way ZPlot accepts the option in case manual plotting is desired.