Return the full vector of radial polynomial coefficient vectors
This allows immediate extraction of the full set of coefficients for
every polynomial from Z(j = 0) to Z(m_max, n_max).

Because they're computed recursively doing this eliminates the need to
call the function multiple times as in the aperture transformations,
allowing for significant performance gains.

I knew the i_max derivation would come in handy eventually, as the get_i
function is what allows the utilization of an indexing scheme for
accessing the corresponding coefficient vector.