WavefrontError: fix an error for n_max = 0
Fitting to zeroth order is essentially mean(OPD) as it extracts only the
piston term. However, because the coefficient vector had its type
declared this would result in an error as the fitting is performed using
a vector instead of a matrix producing a scalar as the output.
So, in order to maintain type stability make sure that if this is done
then the output is a 1-element vector containing the coefficient.