WavefrontError: allow selection of specific fitting orders
This allows specifying exactly which polynomials to use in
reconstructing the wavefront error through an input vector of (m, n)
tuples instead of fitting through the full range.

This information is also encoded in the wavefront error using the fit_to
field, with an empty vector signifying a full fit up to n_max (0:j_max).