WavefrontError: dispatch on the n_max / orders dichotomy
I think orthogonalizing the design using additional barriers in this
manner is generally preferrable as the functionality is mutually
exclusive and this avoids explicit conditional checking, optional
arguments, and unused parameters. This way the construction function can
be accessed directly to obtain the expansion coefficients without having
to pass both n_max and the orders vector of tuples.