Compute inner products more efficiently
For the RadialPolynomials this doesn't seem to make a difference
performance-wise, it's simply a more terse way of expressing it.

For the RMS wavefront error, however, computing the norm squared using
a dot product in this manner is in fact more efficient than broadcasting
and summing. v' * v appears to be implementationally equivalent to
dot(v, v) and properly returns a scalar under vectorial input.