Use the new stack function to construct the phase error fitting matrix
This was recently introduced in Julia 1.9.0 and is more efficient than
the previous solution. It appears to be equivalent to doing:

A = Matrix{T}(undef, length(OPD), length(Zᵢ))
for (i, Z) in pairs(Zᵢ)
    A[:,i] = Z.(ρ, θ)
end

which allocates less memory. Either way, this also does away with the
need for the ternary trick at the end as using a generator with stack
produces a matrix regardless as in the above example.