WavefrontError: adjust the allowed input types
So two things:

A. Make sure we're dealing with floats;
B. Set the requirements light enough so as to allow manipulated arrays
such as transposes, adjoints, reshapes, SubArray views etc.

The output type for the coefficient vector is Float64 as that's what our
polynomials are using so for input Float32 data conversion will be
necessary.

The argument types for the transform aperture functions will remain
fixed at Float64 in accordance with this as it's probably better to keep
everything tight there; that vector has to be in a specific standardized
format anyway.