WavefrontError: implement a matrix input method
The matrix method which existed was only for data aligned such that the
columns represented the pupil positions and optical path difference.
This has been kept as split_data which returns a generator.

The new specialization accepts a matrix of wavefront error data produced
in a polar coordinate system. The first axis of the matrix (the rows) is
assumed to correspond to the angular variable θ while the second axis
(the columns) is assumed to correspond to the radial variable ρ; the
input matrix can simply be transposed if it's the other way around.

The pupil points are simply reconstructed following this assumption.