Explicitly construct the variable ranges using floats
This isn't strictly necessary at the moment since Julia dispatches this
specific method of range(start, stop, length) in such a way that the
start and stop parameters are either promoted to the same type or
converted to floats if they're both integers even if the given length is
such that it would be a unit equivalent (e.g. range(0, 6, 7) yields
floats). Also, in this case the length would have to be 2 for the radial
variable range for any edge case to be a problem.

But, it's better to be explicit in case this changes in the future and
since we're working with floats it's just more consistent.