TransformAperture: anchor the matrix bools outside of the loop
While it's not great that conditions are being checked within the loop
it's better to bind a few vars rather than constantly querying the value
and this is also easier to follow. As stated before the reason why
modularity is difficult to achieve here is because we want as few loops
as possible if multiple transformations are being requested.

Now, because there are so many permutations and so many parameters to
pass compartmentalization becomes convoluted. Since the real valued
scale and the complex valued scale_translate transforms are mutually
exclusive we have n = 3, so for {scaling, rotation, elliptical} there's
2^n - 1 = 7 permutations and likewise for {scale_translation, rotation,
elliptical} another 7 yielding a total of 14.

However, we need at minimum one loop and at most two and there are 3
different permutations for each (scale, scale + rotate, none),
(translate, translate + elliptic, none) so we could follow something
similar to the current pattern with translate/elliptical/transform and
have scale!, rotate!, scale_rotate! functions which accept allocated
arrays and with the use of the identity matrix we can cover the
remaining combinations.

But this is still quite ugly and from the looks of it requires
replication of the substantial S function (four in total) with little to
no performance benefit, so rather than having extra clutter that needs
to be maintained I'm going to leave it as it is for now.

The Unions are small, and it's been benchmarked against more specialized
approaches and the current scheme wins out from the standpoint of
overall simplicity.