TransformAperture: utilize macros to reduce redundancy
Julia allows me to do this so why not. On the one hand it makes the code
easier to read at a high level if you kinda know what it does and you're
just scanning through; on the other, it makes it more difficult to
figure out exactly what's going on because you have to trace out the
macros and realize expressions are being escaped and variables are being
spliced in. Those expressions are long though so it makes sense to
separate the core calculations and reuse them.