WavefrontError: add a phase mapping reverse transform function
For data in the form of associated vectors this will yield the
corresponding phase matrix and coordinate vectors. It's meant to be an
auxiliary helper function in choosing the fitting orders.

If your data is already in the form of a matrix you can easily construct
the ranges given its size and plot it using ZPlot, which when compared
with the various polynomials should assist in formulating an ansatz with
respect to the appropriate n_max or orders.

However, if the data is in canonical form (the fully descriptive triplet
of vectors where the values at each index are linked such that they form
(ρ, θ, OPD) tuples over the pupil) then this is a bit more difficult
given the required input format for ZPlot. That's where this function
comes in, allowing for Zernike.ZPlot(Zernike.map_phase(ρ, θ, OPD)...).