map_phase: don't sort the coordinate vectors, order the reshape properly
So I was going to revert this because it makes quite a few assumptions
about the data and doesn't work for all cases, but it's still useful
when the assumptions hold (which I think should typically be the case)
so instead I'm just fixing it. Obviously the sorting makes sense only if
the OPD is also sorted accordingly and since the original commit didn't
do this the sorting breaks alignment unless the vectors are already
sorted; sorting doesn't actually matter anyway. Also, the matrix was
essentially transposed, so that's been fixed as well.

As for the assumptions, the new comment basically explains it. I was
going to try an approach which inserts NaNs instead in order to
regularize the sampling intervals as Makie will by default assign
transparency to those elements, but this doesn't work well with surface
or our pupil view, so I'm not going to bother. Vectored data like that
is easily plotted using Cartesian coordinates and Makie's stem or
scatter or another plotting package.