TransformAperture: add an unhygienic init macro for the transform loops
I considered this in the initial commit, even implemented it in the
first iteration along with macros for the pre-loop set-up work, but
decided not to go overboard. But, on second thought, since the kernels
use this to me it's clearer while reading through when it's outside of
the functions.