Store the full standardized coefficient vector within the WavefrontError
This is produced on basic output, but was missing when working with
models directly. As the transform functions expect this format for the
input vector it's useful to have it stored as metadata inside the
WavefrontError record.

The field names are still a little confusing and somewhat inconsistent
(e.g. the `a` refers to the subset used for modelling after precision
filtering and doesn't refer to the same `a` field in WavefrontOutput;
this field is instead named `i` within WavefrontError), but as stated
it's probably better to leave it context-dependent rather than having
to go with names like `av` for what is currently `a` and `inds_a` for
what is currently `i`. It makes sense to me, anyway.