Zernike: use evalpoly to evaluate the radial polynomial
So the accuracy is difficult to test due to how the polynomials
fluctuate at high orders along with floating point shenanigans, but
evalpoly is a much faster and more efficient method compared to
broadcasting powers and taking the dot product and since we're calling
this function in loops for the wavefront modelling speed is important.
At high orders both methods will be inaccurate anyway.