ZernikePlot: disable the shading / lighting
This looks better for the types of plots we're generating and it also
brightens up the pupil images since we're using a θ x ρ convention for
our matrices; using a ρ x θ convention or passing invert_normals to
surface! on the other hand causes the reflections and lighting
characteristics to show up on the bird's eye / top-down pupil view so
regardless of which side we use it's best to turn the shading off.