Add the ability to configure plot settings
This design of storing options inside of a mutable struct is preferable
since the plotting function isn't exactly exposed and isn't usually
meant to be called directly. At least this is more convenient than
passing keyword arguments or the like to the main functions wrapping
ZPlot. Also with this method the settings persist.

This way the monitor properties don't need to be extracted on each run,
although this does have the disadvantage that if your main monitor
changes then the sizing settings won't update, but this is probably not
as common as generating multiple plots on the same display; in any case,
I've included a manual resizing mechanism for that purpose.