Implement the plotting using a customized recipe
This allows us to invoke plot methods using our polynomial and wavefront
error functions as well as discretized matrices which is a much more
flexible solution. The plot config is still convenient instead of having
the attributes do the heavy lifting as it allows for overriding the
defaults instead of having to pass the settings as keyword arguments.
Since the outward facing interface wraps the main plotting function
which itself wraps the custom plot and sets the figure and axis this is
ultimately the more sensible approach I think.