ZernikePlot: use `@extractvalue` instead, remove colormap attribute
So, not entirely happy with this design as using a recipe converts all
attributes to Observables, but the attributes we're using shouldn't be
Observables, hence the need to extract. `@extractvalue` exists so that
takes care of the to_value step. colormap is baked into the config so
there's no need to have it as an attribute; initial designs included a
bool read_conf attribute which could toggle whether the config was read
or the attribute was used, but for one attribute this is unnecessary.
Also, there's no need to extract colormap at zplot.

It might be better to not use a recipe at all and just manually dispatch
but this is kind of convenient as it handles the attribute defaults and
a single extraction line is no big deal. I'll commit the redesign next
though as I think it could be much cleaner than what we currently have.
If it doesn't work out I'll just revert it.