cherry-pick b0f7c92 (revert the undo recipe revert)
On second thought this is more flexible, it allows direct access through
keyword arguments and doesn't have to go through any macros or anything,
just good old fashioned functions. The dispatch system is simple, based
on the number of arguments, less restrictive now because we need them
open to Observables.