ZernikePlot: Do away with the recipe system
Okay, so not sure if this is cleaner, but it certainly makes more sense.
It still utilizes the dispatch mechanism to distinguish between ZW type
functions or discretized wavefront approximations so really the only
difference is the need to declare the keyword arguments up front and the
fact that the attributes aren't converted to Observables. The plotconfig
is still used if the keyword arguments aren't passed, but this design is
more flexible in the sense that you could pass them if you want to
deviate from what's set.

Ultimately, the recipe system makes more sense if the attributes are
meant to be dynamic, but this is not the case at the moment. I have a
feeling that I will eventually revert this commit and just build upon
the previous design as visualizing changes to the wavefront error
dynamically could be useful; for example calling zplot with an
Observable{WavefrontError} and varying certain Zernike components to see
how the phase surface changes. Of course we could still use this design
and expand the Union on input argument types to include Observables
instead as only the function needs to vary and not the attributes, but
the recipe approach might still be better for this. Anyway, this is
simply being committed just to have the alternative on record.