Add some WavefrontError constructors
WavefrontErrors can be conceptualized as higher order abstractions in
the sense that they can be thought of as vectors with respect to an
orthogonal basis comprised of the set of Zernike polynomials spanning a
vector space over the unit disk. However, such linear combinations of
Zernike polynomials need not necessarily be wavefront errors so it makes
sense to subtype both Zernike polynomials and wavefront errors as part
of a more general mathematical phenomenon; for lack of a better term
I've decided to call this "Phase".

Now, it can be useful to, in a sense, reify this and grant direct access
to the construction of these instantiations; that is to say,
construction of any form of synthesized, composite Zernike polynomials
without having to go through the fitting process. Since one of our
plotting methods accepts these subtypes (both the pure orthogonal
components of this Phase, and the composite versions) this will allow us
to create and plot any type of these combinations (and with the next
commit, interactively).

All that's really left is to lift the input function on quantization
within the Phase specialization of zernikeplot and we can then vary the
surface plot dynamically by varying our input
Observable{WavefrontError}.